requirements.txt
click==8.0.3
flake8==4.0.1
Flask==2.0.2
flask-sock==0.4.0
greenlet==1.1.2
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.1
simple-websocket==0.4.0
SQLAlchemy==1.4.25
Turbo-Flask==0.6.1
Werkzeug==2.0.2
wsproto==1.0.0
